fix(desktop): render string process ports
